+++ bindings/nostr-sdk-js/src/client/mod.rs
@@ -503,10 +503,7 @@ impl JsClient {
             .map(|id| id.into())
     }
 
-    /// Fetch the newest public key metadata from database and connected relays.
-    ///
-    /// If you only want to consult cached data,
-    /// consider `client.database().profile(PUBKEY)`.
+    /// Fetch the newest public key metadata from relays.
     ///
     /// <https://github.com/nostr-protocol/nips/blob/master/01.md>
     #[wasm_bindgen(js_name = fetchMetadata)]