+++ CHANGELOG.md
@@ -74,6 +74,7 @@
 * ffi: remove `NostrLibrary` struct and keep only `git_hash_version` func ([Yuki Kishimoto])
 * ffi: remove embedded tor client ([Yuki Kishimoto])
 * bindings: cleanup `Relay` methods ([Yuki Kishimoto])
+* bindings: cleanup `NostrDatabase` methods ([Yuki Kishimoto])
 
 ### Deprecated